+++ .gitignore
@@ -5,3 +5,4 @@ cover.out
 .editorconfig
 .idea/
 bash_unit
+dist/qemu-*
